moving Navbar to components folder
