+++ src/App.vue
@@ -4,7 +4,7 @@
 </template>
 
 <script>
-import Navbar from "@/views/Navbar.vue";
+import Navbar from "@/components/Navbar.vue";
 export default {
   name: "App",
   components: { Navbar },
